+++ examples/v2/order/OrderCreate.php
@@ -18,6 +18,9 @@ $order = array();
 
 $order['continueUrl'] = 'http://localhost/examples/v2/order/ContinueUrl.php';
 $order['notifyUrl'] = 'http://localhost/examples/v2/order/OrderNotify.php';
+$order['completeUrl'] = 'http://localhost/examples/layout/success.php';
+$order['cancelUrl'] = 'http://localhost/examples/layout/error.php';
+
 $order['customerIp'] = '127.0.0.1';
 $order['merchantPosId'] = OpenPayU_Configuration::getMerchantPosId();
 $order['description'] = 'New order';
@@ -55,21 +58,35 @@ $response = OpenPayU_Order::create($order);
         <h1>Create Order - OpenPayU v2</h1>
     </div>
     <h1>Request</h1>
+
     <div id="unregisteredCardData">
         <?php var_dump($order); ?>
     </div>
 
     <table class="table table-hover table-bordered">
         <thead>
-        <tr><th colspan="2">Important data from response</th></tr>
+        <tr>
+            <th colspan="2">Important data from response</th>
+        </tr>
         </thead>
         <tbody>
-        <tr><td>Order status</td><td><?=$response->getStatus()?></td></tr>
-        <tr><td>Order id</td><td><?=$response->getResponse()->orderId?></td></tr>
-        <tr><td>Redirect Uri</td><td><a href="<?=$response->getResponse()->redirectUri?>"><?=$response->getResponse()->redirectUri?></a></td></tr>
+        <tr>
+            <td>Order status</td>
+            <td><?= $response->getStatus() ?></td>
+        </tr>
+        <tr>
+            <td>Order id</td>
+            <td><?= $response->getResponse()->orderId ?></td>
+        </tr>
+        <tr>
+            <td>Redirect Uri</td>
+            <td><a href="<?= $response->getResponse()->redirectUri ?>"><?= $response->getResponse()->redirectUri ?></a>
+            </td>
+        </tr>
         </tbody>
     </table>
     <h1>Response</h1>
+
     <div id="unregisteredCardData">
         <?php var_dump($response); ?>
     </div>