+++ examples/v2/order/OrderRetrieve.php
@@ -14,7 +14,6 @@
 require_once realpath(dirname(__FILE__)) . '/../../../lib/openpayu.php';
 require_once realpath(dirname(__FILE__)) . '/../../config.php';
 
-
 ?>
 <!doctype html>
 <html lang="en-US">